requirements.txt
torch>=2.2
torchvision>=0.17
einops>=0.7
open_clip_torch>=2.26
transformers>=4.44
accelerate>=0.33
peft>=0.12
numpy
pillow
tqdm
matplotlib

